Move helpers page from Locations reporting page
