+++ .env.sample
@@ -95,5 +95,5 @@ REACT_APP_NAVBAR_BRAND_IMG_SRC=https://github.com/onaio/reveal-frontend/raw/mast
 REACT_APP_REVEAL_BRAND_IMG_SRC=https://github.com/onaio/reveal-frontend/raw/master/src/assets/images/logo.png
 
 #use school label
-REACT_APP_SHOW_SCHOOL_REPORT_LABEL=true
+REACT_APP_SHOW_MDA_SCHOOL_REPORT_LABEL=true
