+++ src/configs/__mocks__/env.ts
@@ -155,4 +155,4 @@ export const ENABLE_MDA_POINT = true;
 
 export const SUPERSET_MDA_POINT_LOCATION_REPORT_DATA_SLICE = '01';
 
-export const SHOW_SCHOOL_REPORT_LABEL = true;
+export const SHOW_MDA_SCHOOL_REPORT_LABEL = true;
