+++ src/configs/env.ts
@@ -312,5 +312,6 @@ export type NAVBAR_BRAND_IMG_SRC = typeof NAVBAR_BRAND_IMG_SRC;
 export const REVEAL_BRAND_IMG_SRC = process.env.REACT_APP_REVEAL_BRAND_IMG_SRC;
 export type REVEAL_BRAND_IMG_SRC = typeof REVEAL_BRAND_IMG_SRC;
 
-export const SHOW_SCHOOL_REPORT_LABEL = process.env.REACT_APP_SHOW_SCHOOL_REPORT_LABEL === 'true';
-export type SHOW_SCHOOL_REPORT_LABEL = typeof SHOW_SCHOOL_REPORT_LABEL;
+export const SHOW_MDA_SCHOOL_REPORT_LABEL =
+  process.env.REACT_APP_SHOW_MDA_SCHOOL_REPORT_LABEL === 'true';
+export type SHOW_MDA_SCHOOL_REPORT_LABEL = typeof SHOW_MDA_SCHOOL_REPORT_LABEL;
